delete components module and move files
