+++ .idea/gradle.xml
@@ -10,7 +10,6 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/components" />
             <option value="$PROJECT_DIR$/quickedit" />
           </set>
         </option>
