--- components/src/main/java/com/abizer_r/components/util/ColorUtils.kt
@@ -1,20 +1,0 @@
-package com.abizer_r.components.util
-
-import androidx.compose.ui.graphics.Color
-
-object ColorUtils {
-
-    val defaultColorList = arrayListOf(
-        Color.White,
-        Color.Black,
-        Color.Red,
-        Color.Green,
-        Color.Blue,
-        Color.Yellow,
-        Color.Cyan,
-        Color.Magenta,
-        Color.LightGray,
-        Color.Gray,
-        Color.DarkGray,
-    )
-}
