+++ quickedit/build.gradle.kts
@@ -42,8 +42,6 @@ android {
 
 dependencies {
 
-    implementation(project(":components"))
-
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.activity.compose)
     implementation(libs.androidx.appcompat)
