+++ quickedit/src/main/java/com/abizer_r/quickedit/ui/cropMode/CropperScreen.kt
@@ -34,9 +34,9 @@ import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.constraintlayout.compose.Dimension
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.lifecycleScope
-import com.abizer_r.components.R
-import com.abizer_r.components.theme.QuickEditTheme
-import com.abizer_r.components.util.defaultErrorToast
+import com.abizer_r.quickedit.R
+import com.abizer_r.quickedit.theme.QuickEditTheme
+import com.abizer_r.quickedit.utils.defaultErrorToast
 import com.abizer_r.quickedit.ui.common.AnimatedToolbarContainer
 import com.abizer_r.quickedit.ui.common.bottomToolbarModifier
 import com.abizer_r.quickedit.ui.common.topToolbarModifier
@@ -229,7 +229,7 @@ fun PreviewEditorScreen() {
             CropperScreen(
                 animatedVisibilityScope = it,
                 immutableBitmap = ImmutableBitmap(
-                    ImageBitmap.imageResource(id = R.drawable.placeholder_image_1).asAndroidBitmap()
+                    ImageBitmap.imageResource(id = R.drawable.placeholder_image_3).asAndroidBitmap()
                 ),
                 onDoneClicked = {},
                 onBackPressed = {}
