+++ quickedit/src/main/java/com/abizer_r/quickedit/ui/textMode/TextModeScreen.kt
@@ -33,9 +33,9 @@ import androidx.constraintlayout.compose.Dimension
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.lifecycleScope
-import com.abizer_r.components.util.ImmutableList
+import com.abizer_r.quickedit.utils.ImmutableList
 import com.abizer_r.quickedit.utils.textMode.blurBackground.BlurBitmapBackground
-import com.abizer_r.components.util.defaultErrorToast
+import com.abizer_r.quickedit.utils.defaultErrorToast
 import com.abizer_r.quickedit.ui.common.AnimatedToolbarContainer
 import com.abizer_r.quickedit.ui.common.bottomToolbarModifier
 import com.abizer_r.quickedit.ui.common.topToolbarModifier
