+++ quickedit/src/main/java/com/abizer_r/quickedit/ui/textMode/TextModeState.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.sp
-import com.abizer_r.components.util.ColorUtils
+import com.abizer_r.quickedit.utils.ColorUtils
 import com.abizer_r.quickedit.ui.transformableViews.base.TransformableBoxState
 
 data class TextModeState(
