+++ quickedit/src/main/java/com/abizer_r/quickedit/utils/CommonExtensions.kt
@@ -1,11 +1,14 @@
 package com.abizer_r.quickedit.utils
 
+import android.content.Context
+import android.widget.Toast
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.AnimatedVisibilityScope
 import androidx.compose.animation.ExperimentalSharedTransitionApi
 import androidx.compose.animation.SharedTransitionLayout
 import androidx.compose.animation.SharedTransitionScope
 import androidx.compose.runtime.Composable
+import com.abizer_r.quickedit.R
 
 val Any.TAG: String
     get() {
@@ -23,4 +26,12 @@ fun SharedTransitionPreviewExtension(
             content(this)
         }
     }
+}
+
+fun Context.toast(message: String) {
+    Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
+}
+
+fun Context.defaultErrorToast() {
+    Toast.makeText(this, this.getString(R.string.something_went_wrong), Toast.LENGTH_SHORT).show()
 }
