+++ quickedit/src/main/java/com/abizer_r/quickedit/utils/textMode/colorList/ColorListFullWidth.kt
@@ -14,10 +14,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.abizer_r.components.theme.QuickEditTheme
-import com.abizer_r.components.theme.ToolBarBackgroundColor
-import com.abizer_r.components.util.ColorUtils
-import com.abizer_r.components.util.ImmutableList
+import com.abizer_r.quickedit.theme.QuickEditTheme
+import com.abizer_r.quickedit.theme.ToolBarBackgroundColor
+import com.abizer_r.quickedit.utils.ColorUtils
+import com.abizer_r.quickedit.utils.ImmutableList
 
 /**
  * The total size of this item is the adding of [itemSize] and [selectedBorderWidth]
