+++ quickedit/src/main/res/drawable/ic_color_picker.png
