+++ quickedit/src/main/res/drawable/ic_eraser.xml
