+++ quickedit/src/main/res/drawable/ic_stylus_note.xml
