+++ quickedit/src/main/res/values/strings.xml
